requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
Django==4.0.6
django-rest-framework==0.1.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
google-auth==2.11.1
idna==3.4
ifaddr==0.2.0
kubernetes==24.2.0
oauthlib==3.2.1
py-eureka-client==0.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.12.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.2.4
urllib3==1.26.12
uuid==1.30
websocket-client==1.4.1
